backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
livekit-api==1.0.3
